requirements.txt
# requirements.txt
fastapi
python-dotenv
uvicorn
pydantic
dependency-injector
pymongo
nats-py>=2.2.0
grpcio
protobuf
